+++ src/main/java/com/welldone/springboot/model/OrderItems.java
@@ -1,9 +1,7 @@
 package com.welldone.springboot.model;
 
-import com.welldone.springboot.model.pk.OrderItemPK;
-import jakarta.persistence.EmbeddedId;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
+
+import jakarta.persistence.*;
 
 import java.io.Serializable;
 import java.util.Objects;
@@ -11,29 +9,38 @@ import java.util.Objects;
 @Entity
 @Table(name = "tb_OrderItems")
 public class OrderItems implements Serializable {
-    @EmbeddedId
-    private OrderItemPK id;
+
+    @ManyToOne
+    @JoinColumn(name = "order_id")
+    private Order order;
+    @ManyToOne
+    @JoinColumn(name = "product_id")
+    private Product product;
     private Integer qnt;
     private Double price;
 
-    public Order getOrder(){
-        return id.getOrder();
-    }
-    public void setOrder(Order order){
-        id.setOrder(order);
-    }
-    public Product getProduct(Product product){
-        return id.getProduct();
-    }
-    public void setProduct(Product product){
-        id.setProduct(product);
+    public Order getOrder() {
+        return order;
     }
 
-    public OrderItems(){}
+    public void setOrder(Order order) {
+        this.order = order;
+    }
+
+    public Product getProduct() {
+        return product;
+    }
+
+    public void setProduct(Product product) {
+        this.product = product;
+    }
+
+    public OrderItems() {
+    }
 
     public OrderItems(Order order, Product product, Integer qnt, Double price) {
-        id.setOrder(order);
-        id.setProduct(product);
+        this.order = order;
+        this.product = product;
         this.qnt = qnt;
         this.price = price;
     }
@@ -58,11 +65,14 @@ public class OrderItems implements Serializable {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (!(o instanceof OrderItems that)) return false;
-        return Objects.equals(id, that.id);
+        return Objects.equals(getOrder(), that.getOrder()) && Objects.equals(getProduct(), that.getProduct());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id);
+        return Objects.hash(getOrder(), getProduct());
     }
 }
+
+
+
