--- src/main/java/com/welldone/springboot/model/pk/OrderItemPK.java
@@ -1,48 +1,0 @@
-package com.welldone.springboot.model.pk;
-
-import com.welldone.springboot.model.Order;
-import com.welldone.springboot.model.Product;
-import jakarta.persistence.Embeddable;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-@Embeddable
-public class OrderItemPK implements Serializable {
-    @ManyToOne
-    @JoinColumn(name = "order_id")
-    private Order order;
-    @ManyToOne
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    public Order getOrder() {
-        return order;
-    }
-
-    public void setOrder(Order order) {
-        this.order = order;
-    }
-
-    public Product getProduct() {
-        return product;
-    }
-
-    public void setProduct(Product product) {
-        this.product = product;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof OrderItemPK that)) return false;
-        return Objects.equals(getOrder(), that.getOrder()) && Objects.equals(getProduct(), that.getProduct());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(getOrder(), getProduct());
-    }
-}
